requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.5
django-restframework==0.0.1
djangorestframework==3.12.4
Faker==8.10.1
pycodestyle==2.7.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
